+++ CHANGELOG.md
@@ -5,13 +5,14 @@
 **Added**
 
 -   Support for Laravel 10 [#349](https://github.com/auth0/laravel-auth0/pull/349)
+-   New `Auth0\Laravel\Traits\Imposter` trait to allow for easier testing.
 
 **Changed**
 The following changes have no effect on the external API of this package, but may affect internal usage.
 
 -   `StateInstance` concept has been replaced by new `Credentials` entity.
 -   `Guard` updated to use new `Credentials` entity as primary internal storage for user data.
--
+-   `Auth0\Laravel\Traits\ActingAsAuth0User` was updated to use new`Credentials` entity.
 
 Thanks to our contributors for this release: [taida957789](https://github.com/taida957789)